update unit test project
