--- AstronomyPictureOfTheDay.Tests/packages.config
@@ -1,13 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="Castle.Core" version="4.4.1" targetFramework="net47" />
-  <package id="FakeItEasy" version="6.2.1" targetFramework="net47" />
-  <package id="Newtonsoft.Json" version="12.0.3" targetFramework="net47" />
-  <package id="xunit" version="2.4.1" targetFramework="net47" />
-  <package id="xunit.abstractions" version="2.0.3" targetFramework="net47" />
-  <package id="xunit.analyzers" version="0.10.0" targetFramework="net47" />
-  <package id="xunit.assert" version="2.4.1" targetFramework="net47" />
-  <package id="xunit.core" version="2.4.1" targetFramework="net47" />
-  <package id="xunit.extensibility.core" version="2.4.1" targetFramework="net47" />
-  <package id="xunit.extensibility.execution" version="2.4.1" targetFramework="net47" />
-</packages>
