+++ AstronomyPictureOfTheDay.sln
@@ -4,9 +4,9 @@ VisualStudioVersion = 16.0.30523.141
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AstronomyPictureOfTheDay", "AstronomyPictureOfTheDay\AstronomyPictureOfTheDay.csproj", "{3EAECD48-CB4D-43A2-97C3-490704E7D473}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AstronomyPictureOfTheDay.Tests", "AstronomyPictureOfTheDay.Tests\AstronomyPictureOfTheDay.Tests.csproj", "{246765B1-7A32-41FE-8BBD-E0D413A7DE09}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AstronomyPictureOfTheDay.Sample.WPFCore", "AstronomyPictureOfTheDay.Sample.WPFCore\AstronomyPictureOfTheDay.Sample.WPFCore.csproj", "{295B887D-62A3-4DB9-B2A4-C22A9EFB4CB1}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AstronomyPictureOfTheDay.Sample.WPFCore", "AstronomyPictureOfTheDay.Sample.WPFCore\AstronomyPictureOfTheDay.Sample.WPFCore.csproj", "{295B887D-62A3-4DB9-B2A4-C22A9EFB4CB1}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AstronomyPictureOfTheDay.Xunit.Tests", "AstronomyPictureOfTheDay.Xunit.Tests\AstronomyPictureOfTheDay.Xunit.Tests.csproj", "{8DBC95A7-32F8-4071-AF4F-D0DC7963E912}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -30,18 +30,6 @@ Global
 		{3EAECD48-CB4D-43A2-97C3-490704E7D473}.Release|iPhone.Build.0 = Release|Any CPU
 		{3EAECD48-CB4D-43A2-97C3-490704E7D473}.Release|iPhoneSimulator.ActiveCfg = Release|Any CPU
 		{3EAECD48-CB4D-43A2-97C3-490704E7D473}.Release|iPhoneSimulator.Build.0 = Release|Any CPU
-		{246765B1-7A32-41FE-8BBD-E0D413A7DE09}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{246765B1-7A32-41FE-8BBD-E0D413A7DE09}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{246765B1-7A32-41FE-8BBD-E0D413A7DE09}.Debug|iPhone.ActiveCfg = Debug|Any CPU
-		{246765B1-7A32-41FE-8BBD-E0D413A7DE09}.Debug|iPhone.Build.0 = Debug|Any CPU
-		{246765B1-7A32-41FE-8BBD-E0D413A7DE09}.Debug|iPhoneSimulator.ActiveCfg = Debug|Any CPU
-		{246765B1-7A32-41FE-8BBD-E0D413A7DE09}.Debug|iPhoneSimulator.Build.0 = Debug|Any CPU
-		{246765B1-7A32-41FE-8BBD-E0D413A7DE09}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{246765B1-7A32-41FE-8BBD-E0D413A7DE09}.Release|Any CPU.Build.0 = Release|Any CPU
-		{246765B1-7A32-41FE-8BBD-E0D413A7DE09}.Release|iPhone.ActiveCfg = Release|Any CPU
-		{246765B1-7A32-41FE-8BBD-E0D413A7DE09}.Release|iPhone.Build.0 = Release|Any CPU
-		{246765B1-7A32-41FE-8BBD-E0D413A7DE09}.Release|iPhoneSimulator.ActiveCfg = Release|Any CPU
-		{246765B1-7A32-41FE-8BBD-E0D413A7DE09}.Release|iPhoneSimulator.Build.0 = Release|Any CPU
 		{295B887D-62A3-4DB9-B2A4-C22A9EFB4CB1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{295B887D-62A3-4DB9-B2A4-C22A9EFB4CB1}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{295B887D-62A3-4DB9-B2A4-C22A9EFB4CB1}.Debug|iPhone.ActiveCfg = Debug|Any CPU
@@ -54,6 +42,18 @@ Global
 		{295B887D-62A3-4DB9-B2A4-C22A9EFB4CB1}.Release|iPhone.Build.0 = Release|Any CPU
 		{295B887D-62A3-4DB9-B2A4-C22A9EFB4CB1}.Release|iPhoneSimulator.ActiveCfg = Release|Any CPU
 		{295B887D-62A3-4DB9-B2A4-C22A9EFB4CB1}.Release|iPhoneSimulator.Build.0 = Release|Any CPU
+		{8DBC95A7-32F8-4071-AF4F-D0DC7963E912}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{8DBC95A7-32F8-4071-AF4F-D0DC7963E912}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{8DBC95A7-32F8-4071-AF4F-D0DC7963E912}.Debug|iPhone.ActiveCfg = Debug|Any CPU
+		{8DBC95A7-32F8-4071-AF4F-D0DC7963E912}.Debug|iPhone.Build.0 = Debug|Any CPU
+		{8DBC95A7-32F8-4071-AF4F-D0DC7963E912}.Debug|iPhoneSimulator.ActiveCfg = Debug|Any CPU
+		{8DBC95A7-32F8-4071-AF4F-D0DC7963E912}.Debug|iPhoneSimulator.Build.0 = Debug|Any CPU
+		{8DBC95A7-32F8-4071-AF4F-D0DC7963E912}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{8DBC95A7-32F8-4071-AF4F-D0DC7963E912}.Release|Any CPU.Build.0 = Release|Any CPU
+		{8DBC95A7-32F8-4071-AF4F-D0DC7963E912}.Release|iPhone.ActiveCfg = Release|Any CPU
+		{8DBC95A7-32F8-4071-AF4F-D0DC7963E912}.Release|iPhone.Build.0 = Release|Any CPU
+		{8DBC95A7-32F8-4071-AF4F-D0DC7963E912}.Release|iPhoneSimulator.ActiveCfg = Release|Any CPU
+		{8DBC95A7-32F8-4071-AF4F-D0DC7963E912}.Release|iPhoneSimulator.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
